osij_backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
python-dotenv>=1.0.0
django-cors-headers>=4.0.0
Pillow>=10.0.0
djangorestframework-simplejwt>=5.3.0
django-paypal>=1.2.5
paypalrestsdk>=1.13.3